Make an example localsettings.js file.
The current installation instructions suggest reading ParserService.js
to create localsettings.js. Adding this example file should make
the installation a bit easier - the user will simply have to copy
the example file to the target name and edit it.

Also fixed the line ending in the .gitignore entry for localsettings.js.

Change-Id: Ia4f16ac636cbd3473837f439c699e5e3e9950e9e